Change "uint" -> size_t
Using size_t fixes compiler errors under Windows.
(MSVC doesn't like "uint" as it is not a "real" type.)

+++ src/SweepComponentProcessor.h
@@ -167,7 +167,7 @@ public:
     return {};
   }
 
-  std::vector<float> getFrequencyResponse(uint numbins)
+  std::vector<float> getFrequencyResponse(size_t numbins)
   {
     jassert(inputBuffer);
     jassert(sweep);
