+++ src/fft.h
@@ -21,7 +21,7 @@ std::vector<float> dft_magnitude(std::vector<float> input);
 std::vector<float> dft_magnitude_db(std::vector<float> input);
 std::vector<float> dft_magnitude_with_log_bins(const std::vector<float>& input,
                                                float sampleRate,
-                                               uint numbins);
+                                               size_t numbins);
 std::vector<float> dft_log_bins(size_t num_samples, float f_low, float f_high);
 std::vector<float> dft_lin_bins(float fs, size_t numSamples);
 RealVector dft_phase(RealVector input);
@@ -32,5 +32,5 @@ RealVector convolve(RealVector a, RealVector b);
 std::vector<float> convolve(const std::vector<float>& a,
                             const std::vector<float>& b);
 
-std::vector<uint> map_log_to_lin_bins(const std::vector<float>& lin_bins,
-                                      const std::vector<float>& log_bins);
+std::vector<size_t> map_log_to_lin_bins(const std::vector<float>& lin_bins,
+                                        const std::vector<float>& log_bins);
